requirements_base.txt
Django>=2.1,<2.2
django-datatables-view>=1.14.0,<1.15.0
django-dynamic-preferences>=1.6,<1.7
django-widget-tweaks>=1.4.1,<1.5
persisting-theory>=0.2.1,<0.3
pytz==2018.4
six>=1.11.0,<1.12
django-author>=1.0.2,<1.1
Pillow>=5.2.0,<5.3
pyreportjasper>=1.0,<1.1
django-sendfile2>=0.4,<0.5